add offline_sign example and setup global build system
