+++ deps/compact25519/c25519/edsign.c
