+++ examples/offlline_sign/constants.h
@@ -3,6 +3,7 @@
 
 #include <stdint.h>
 
+const char* message_hex;
 extern const uint8_t private_key[32];
 
 #endif
