--- src/.gitignore
@@ -1,2 +1,0 @@
-test_sign.out
-constants.c
